requirements.txt
python-dateutil
python-magic>=0.4.6
pytz>=2014.7
Django==1.11
django-markdown-deux
django-mathfilters
six>=1.8.0
Pillow
tornado==4.4
django-imagekit
django-mailjet
captcha
mkdocs
pydenticon
djangorestframework
django-filter
django-rest-swagger
requests-html
pyzmq
circus
web3
